fixed an issue with newly created Team IDs
there was an issue where the last generated team would have the same ID as the first or an ID of 0(only in the program, not in the database), so the program would display employees either not assigned to a team or that were assigned to the first team.

This only occured on the local database and not the school one. hopefully, it's all been fixed now